+++ src/main/java/com/tiernebre/web/routes/LoginRoutes.java
@@ -1,20 +1,20 @@
-package com.tiernebre.web.routes.api;
+package com.tiernebre.web.routes;
 
 import static io.javalin.apibuilder.ApiBuilder.*;
 
 import com.tiernebre.web.controllers.authentication.AuthenticationController;
 import io.javalin.apibuilder.EndpointGroup;
 
-public final class AuthenticationRoutes implements EndpointGroup {
+public final class LoginRoutes implements EndpointGroup {
 
   private final AuthenticationController controller;
 
-  public AuthenticationRoutes(AuthenticationController controller) {
+  public LoginRoutes(AuthenticationController controller) {
     this.controller = controller;
   }
 
   @Override
   public void addEndpoints() {
-    post("authenticate", controller::handle);
+    post("login", controller);
   }
 }
