+++ src/main/java/com/tiernebre/web/routes/api/ApiRoutes.java
@@ -6,21 +6,15 @@ import io.javalin.apibuilder.EndpointGroup;
 
 public final class ApiRoutes implements EndpointGroup {
 
-  private final AuthenticationRoutes authenticationRoutes;
   private final HealthRoutes healthRoutes;
 
-  public ApiRoutes(
-    AuthenticationRoutes authenticationRoutes,
-    HealthRoutes healthRoutes
-  ) {
-    this.authenticationRoutes = authenticationRoutes;
+  public ApiRoutes(HealthRoutes healthRoutes) {
     this.healthRoutes = healthRoutes;
   }
 
   @Override
   public void addEndpoints() {
     path("/api", () -> {
-      authenticationRoutes.addEndpoints();
       healthRoutes.addEndpoints();
     });
   }
